refactor: BM 상세 조회 기능 처리 위치 수정
