+++ src/main/java/com/pitchain/service/BmCommandService.java
@@ -1,11 +1,14 @@
 package com.pitchain.service;
 
+import com.pitchain.common.apiPayload.ErrorStatus;
 import com.pitchain.common.constant.S3UploadTarget;
 import com.pitchain.common.exception.GeneralException;
+import com.pitchain.dto.BmWithScrapDto;
 import com.pitchain.dto.req.BmCreateReq;
 import com.pitchain.dto.req.BmUpdateReq;
-import com.pitchain.entity.Bm;
-import com.pitchain.entity.Company;
+import com.pitchain.dto.res.BmDetailRes;
+import com.pitchain.dto.res.PtImgRes;
+import com.pitchain.entity.*;
 import com.pitchain.jwt.MemberDetails;
 import com.pitchain.repository.BmRepository;
 import com.pitchain.repository.EntityFacade;
@@ -23,6 +26,9 @@ public class BmCommandService {
     private final EntityFacade entityFacade;
     private final BmSubcategoryService bmSubcategoryService;
     private final PtImgService ptImgService;
+    private final BmScrapService bmScrapService;
+    private final SpService spService;
+    private final MyBmHistoryService myBmHistoryService;
     private final S3Service s3Service;
 
     @Transactional
@@ -32,6 +38,41 @@ public class BmCommandService {
         Bm newBm = bmCreateReq.createBm(company, descImgKey);
         bmRepository.save(newBm);
         bmSubcategoryService.saveAll(newBm.getId(), bmCreateReq.subCategories());
+    }
+
+    @Transactional
+    public BmDetailRes getBmDetail(MemberDetails memberDetails, Long bmId) {
+        Member member = entityFacade.getMember(memberDetails.id());
+
+        BmWithScrapDto bmWithScrapDto = bmRepository.getBmWithScrapDto(member.getId(), bmId)
+                .orElseThrow(() -> new GeneralException(ErrorStatus.BM_NOT_FOUND));
+        Bm bm = bmWithScrapDto.getBm();
+
+        long scrapCnt = bmScrapService.countByBm(bm.getId());
+
+        List<PtImg> ptImgs = ptImgService.getPtImgsByBmId(bm.getId());
+        List<PtImgRes> ptImgResList = ptImgs.stream()
+                .map(ptImg -> PtImgRes.createRes(ptImg.getSerialNum(), ptImg.getImgKey()))
+                .toList();
+
+        List<BmSubCategory> bmSubCategories = bmSubcategoryService.getBmSubCategoryByBmId(bmId);
+        List<String> subcategoriesInKorean = convertToKoreanName(bmSubCategories);
+
+        List<Sp> sps = spService.getSpsByBmId(bm.getId());
+        List<String> spURLs = sps.stream()
+                .map(Sp::getSpKey)
+                .toList();
+
+        myBmHistoryService.saveMyBmHistory(member, bm);
+
+        return BmDetailRes.createRes(bmWithScrapDto, ptImgResList, scrapCnt, subcategoriesInKorean, spURLs);
+    }
+
+    private static List<String> convertToKoreanName(List<BmSubCategory> bmSubCategories) {
+        List<String> subcategoriesInKorean = bmSubCategories.stream()
+                .map(bmSubCategory -> bmSubCategory.getSubCategory().getKoreanName())
+                .toList();
+        return subcategoriesInKorean;
     }
 
     @Transactional
